Split main.rs into library and cli parts
